iso7185prt1702a
1. Added active variant checking. In source order:

a. The number of output instructions was increased by 2.

b. Each field now tracks if it is part of a variant and what variant that is, down to which value variant it is. Also keeps track of the id entry for the tagfield for the variant, this is necessary since that holds the address of the tagfield.

c. Added new option for variant checking.

d. checkvrnt() function added to select that generates the variant checking code.

2. Added check to function assign check for undefined function pointer. This prevents double errors due to malformed functions.