iso7185prt1847
Added check for exponent to large for integer. The original real number constant handling just verified the digits and passed them on to the back end. We now check the exponent fits in integer before passing it on. The mantissa won't overflow, but the exponent overflows when it reaches pint. It introduces a machine dependency, the integer size in pcom, but that is already present in integers. In any case, the exponent is usually less critical, and often is smaller than integer in size.