iso7185pat.pas
Fix for multiple variants for a set of constants. When multiple tagfield constants evaluate to a single variant, and we are generating active variant checks, we have to check all possible constants associated with the variant.
This could have been done with a set, since the values must be contiguous. However, a tagfield can have any base, thus 10000..10010 is a valid tagfield as long as the included constants are accounted for (that is a ISO 7185 requirement). Thus the reasonable solution was to generate a series of compares. I didn't want to generate temp labels and jumps, so did this by accumulating the comparisons in a boolean variable kept on stack.

pcom.pas

1. The new instructions cks and cke were generated. These start and end a variant compare series.

2. A caslst field was added to the variant record field structure. This is then processed to become a list of variants that all use the same set of constants in fieldlist(). You will note that subvar actually does the same thing initially, but is changed to become a pointer to the field list shared by all cases, which the compiler needs elsewhere. So we just copied the same method used to create that list initially.

3. Our friend checkvrnt() now generates the start and end, then iterates through the list of variant cases and generates a series of compares.

pint.pas

1. The new instructions cks and cke are added. They are typeless.

2. The way the cks-ckv-cke set now work is that cks pushes a boolean zero on the stack. Then each ckv checks the stacked variable against the variant constant given in the instruction, and or's the result of that into the stacked boolean. At cke time, the boolean will be true if any of the case constants were found, and thus if it is 0, the check fails with error.