Quite a bit of fixes to pass active variant checks. A common theme was setting the tagfield AFTER assigning the desired variant. Others were more difficult, such as assigning to the store[] array using two different variants in the same statement. I included buffer variables to accomplish that one.
P2 is very interesting because it is very short, and that the designers attempted to run the interpretation from a formal tagged variant. P4 gave this up and made it an undiscriminated type.

You have to wonder how much effort should be expended on correcting the program to meet such requirements.