iso7185prt1801 iso7185prt1802 iso7185prt1803 iso7185prt1804 iso7185prt1805 iso7185prt1806 iso7185prt1807
Instituting "for" loop index threat checks, also started identifier reference checking (incomplete).

The for checks come in terms of a threat flag and a for loop nesting counter. These go into ids with klass vars. The threat flag is set whenever a threaten event happens in a NESTED routine (static level). The for counter simply counts nesting level if the variable is used for a for loop index.

When the for loop is entered, and any threat from a nested routine has been registered against the variable, an error is output. Note that threats in the same routine don't count if they are not within the for loop.

Any nesting count > 1 in the for loop count is automatically an error, since the count is kept per id, it means it was reused as a for index in a nested for. We track nesting mainly to keep from flagging further errors later in the same block. It WILL output multiple errors for multiple nested fors using the same variable.

Inside the for loop, as indicated by the for nest counter, any threats to the variable result in an immediate error.

Modifications:

1. threat and forcnt added to id record. Refer added for reference check later. threat was moved to vars only since that is the only place it is relivant.

2. for threat error added.

3. Refer flag added to labels as well.

4. newlabel inits the refer flag false on labels.

5. All of the declaration routines init refer false. Vardeclaration sets refer, threat and forcount false.

6. expression() and all its subroutines get a threaten flag. This determines if threats should be processed to a variable if that is parsed in factor (starting with id). It was necessary to pass this information down in expression() because callnonstandard() processes all parameters as expressions, then outputs an error on var parameters if the expression was not a strict var. We don't know what leading id expression() processed, so it was either pass the threat request down, or pass the leader id up.

7. variable() in call() gets a threaten parameter, and processes the threat as described in the front of this commit. This then is called with threat true during the processing of read[ln]() parameters.

8. In callnonstandard(), we determine the var status of the parameter and pass it down as threaten to expression.

9. In expression, we process the threat as for variable() when an id leader is seen. This is the standard beginning of a var parameter.

10. In assignment(), we detect that the left hand is a variable and process the threat as above.

11. In gotostatement, we flag refer. This will be used later.

12. In forstatement, we process the index var as described at the start of this commit.