iso7185prt1848
Extended new packed VAR parameter check to WITH statement. The WITH statement passes the packed status of the record accessed in the display as a new field packed. This was simply a convenient place to put it. When selector() sees a raw field, which it would never encounter outside of a with statement, it uses the packed status in the display to start the chain of packed tests in the variable construction.

Note that the packed status of a gattr at the end of an expression is a concatenation, or "or", of all the packed accesses that formed the reference. This is because any packed derivation means the resulting access is to a packed element.