Several missing undefined handling issues:
1. The routine compare works direct on storeop[]. Several opcodes use this, but manually compare the last byte, so those needed moding as well.

2. The mov instruction copies value parameters, and thus needed to transfer the def bits as well.

3. pack and unpack (pck/upk) needed to transfer the def bits from source to definition, but they are stated to check the source for defined (see iso7185prt1727 and iso7185prt1730). Thus these bits are defined by definition.

The undefined fixes bring up a lot of issues. In the mov, pck and upk we have the possibility that an undefined or X value can travel to another address! This is not real with pck and upk, because the ISO 7185 also requires them to be defined at the time of transfer, not later (even though this would be logically correct). In mov, which is used to create value parameters, it is appropriate to fail the undefined later, on use, and not on the transfer to the value parameter. This at least bonds the error to the access of the variable that is undefined. Otherwise, the errors would occur in a shower at the start of the called routine.

Note P5's unique handling of value parameters here. Not copied ONTO the stack but rather copied INSIDE the routine. This can have implications for P6.