Observed that a case such as:
packed [1..10] of char;

or missing type will lead properly into typ() but then exit with error and not set the fsp return parameter. In normal bad type cases, this gets set to nil, which is the normal error fallback for bad types. This is then extended to cases above.

Note that this does not solve bad handling of the nil type case in general. There is another ticket for this.