Fixed undefined issue in pint.pas
The original idea was that only the first byte of a multibyte object like integer or real needed to be defined, since that was the only part that would be checked. The logic error here was that there are a few places where things get moved around without regard to their lengths, and pack/unpack is one of them.

The fix was to set all definition bits in a putxxx() object.