iso7185prt1720 iso7185prt1721
Tagged new and dispose compare fix. We define a new version of both new and dispose, and for each we pushdown a list of the tagfield constants listed in the call onto the stack, with the number of tagfields stacked at the top. The alternate version of new allocates extra space for the tag constant list and places it under the dynamic variable in "negative space", that is, under the normal program data. The number of tags is at the top so that no matter how many tags exist, we can find out the total tag list length by looking one step behind the pointer.

In dispose, it checks that both the number of tags and their values match from the stacked values to the values stored in the dynamic variable, and gives an error if not equal. Thus new/dispose mismatches are handled completely.

Note that this scheme as applicability on other checks, for example iso7185prt1719, changing a tagfield used to allocate the dynamic.