iso7185prt1719
First cut at checks for assignment to tagfield used to allocate record. Added new instruction CTA. This instruction is added after we have a pair on stack of the address to be assigned on stack, followed by the new tagfield value. The address will have come from a pointer to a record with a tagfield, after it has been offset to the field. It does not matter how many tagfields deep it is, they are gathered into a single offset value in idplmt. CTA takes the offset and the level of the tagfield, which is how many tagfields deep it is. CTA then takes the pointer, subtracts the offset to find the original pointer, then finds the tag in the tag case values array in negative space. If it exists, it is compared with what is being assigned it. If they are not equal, it means this is an attempt to change one of the tagfields used to allocate the dynamic variable.

This needs to be both extensively tested and we need to handle the case where a with is used to establish the access.