iso7185prt1851
Set variants undefined after change in activation. A new instruction is created, ivt, that invalidates based on change to tagfield. This instruction expects the address of the tagfield as second on stack, followed by the new value at stack top, the offset from the tagfield to the base of the variant as first parameter, and the maximum size of the variant as the second parameter.

The instruction compares the new tag with the existing tag, if different, the entire variant allocation is set undefined. The check for same tag is required because ISO 7185 only specifies change to what variant is active, not assigns to the tag as setting the new variant undefined. If the tag is undefined, the variant is always set undefined.

1. A new instruction was added, IVT.

2. The varsaddr and varssize variables in identifiers are set from the base and extent of the variant in fieldlist. Note that these variables were added previously, sorry about spreading the fix over two commits.

3. Gen2() gets a new instruction case.

4. Assignment gets a new flag for if a tag assign is to be checked. If so, the load of the destination is forced into load address, and this overrides the direct store optimization. This is necessary to keep IVT relatively simple, there are a lot of store variations.

5. Generate IVT.

6. Change optimized store to address indirect store on tagasc.

7. lsttop was added and set to the pctop before alignment. The aligned pctop was causing undefined access errors in the listing when on. This bug was discovered during testing.

8. lstins was fixed up to correctly list double parameter instructions.

9. Fixed several instructions for correct parameter length.

10. Added IVT assembly.

11. Added IVT instruction.