Issue with detection of var referenced components of packed.
The existing detection of var reference packed just gave error if a var reference is packed. This is not the correct answer, the right answer is a var reference to a COMPONENT of a packed structure. The gattr.packing gives true if the developed reference is either itself packed or is a component of packed. We introduce a new boolean, packcom, which means if the gattr is a component of a packed structure. This flag is identical the packing flag, but delayed by one step when descending into a reference. Thus it says if the subcomponent is the child of a packed structure. We also create that flag in the display along with packing, and save it during a "with" statement.

Besides this general redefinition, several side issues were found. First, selector didn't set packing according to if the top reference was packed. This would cause a bad determination if the vars ref was a packed structure.

Second, pointer dereferencing needed to start several flags over again. It doesn't matter if it is part of a packed structure, it starts out anew as a reference. Note that the pointer could be itself packed.

Third, "with" was storing the actual packed state of the gattr type. Instead it should act as if storing the state of selector, and continuing. Thus it should use the gattr.packing flag.