Correction to var parameter of packed component. The "info" variables carried with gattr of:
gattr.packcom - Indicates object is a component of a packed structure (record or array).
gattr.packing - Indicates object itself is packed or is a component of a packed structure.

This is a tad confusing, especially since the flag of the same name in structure indicates the packing status of just that structure. That was my unfortunate addition, since p4 didn't track packing of structures. The reason you care if not only an object is packed, but is part of a structure that is packed, at any level, is that once a structure is packed, it means all of its components cannot be referenced as var, because they may no longer be addressed at word/byte boundaries. A strict packer could compress those structures down to bits.

So, gattr.packing indicates the object pointed to, or any structure containing it, is packed. Since the var restriction does not care if the object pointed to is packed, but only if it is a component of packed, it means that gattr.packcom is gattr.packing lagged by one level when descending a referernce (selector).

So with that out of the way, the code was calculating this state not from the element being referenced in an array or record, but the array or record itself. This was moved. The other mod was to pointer reference in the same routine, which now specifies the gattr.packing flag directly. Notice that the with statement references gattr, then typtr^, in that order. This means that the packing flag comes from the deepest nested, or typtr^, which was wrong.