Two issues with ckla instruction. First, the dereference flag was being set on assignments. This appears to be a cut and paste error. Second, there is a special encoding done on the chka instruction that needed to also be done on the ckla instruction, its brother.
ckla comes from a time when I thought there needed to be a special bounds check on tagged pointers with the case list packed behind the pointer. This later turned out not be true, but I left the separate instruction on such pointers in case this is needed in the future.