In ivt, always invalidating the variant if the tag is undefined led to overwriting the variant if the record was allocated by a tag constant newspc(). ivt handles both standard variant records and pointers.
The idea of invalidating the variant on undefined tags was to ensure that writes to the variant without setting the tag could create parts of the variant that were defined when a tag change is made. However, this case should be handled by active tag checks.

Thus we remove the feature.