1. Fixed 1848, check for variable reference to packed variable in with statement. The with statement to selector() data pass via gattr was not correct. The key is to treat with r do x ... where x is a field of r as equivalent to the access r.x.
2. Several preliminaries for the undiscriminated tagfields check.

a. Two flags were created for the feature, chkudtc and chkudtf. The reason there are two flags is that one is the "candidate" and one is the "final". Only the final flag actually controls the option. The candidate is controlled by the user in a "option comment". The final flag is copied from the candidate when the "program" symbol is encountered, so the user file must set that BEFORE the program starts.

Now I'll reveal my bias here, I am not comfortable with the user having the ability to flip options on and off anywhere in the program code, and this is going to get changed to a model where it is set before the program run in P6. The undiscriminated tagfield check most certainly needs to be the same across the compile run, so thus the special handling.

b. For fieldlist(), the chkudtf flag overrides the removal of the tagfield in the record allocation. Note that the fact it has no name string for the tag is immaterial. It always had an identifier record, it just didn't allocate it by advancing the address.

c. Selector gets a bit passed to it that indicates if the selector() eval is part of the left side of an assignment.

This is it for now. A problem was uncovered with the ivt instruction, which will be used also to do the undiscriminated tagfield check. Past that, the plan is as follows:

We need to treat a write to a field in a variant as a set of any number of variants nested that enclose it. Thus, the present tagfield checks that are done when descending into a nested series of variants are changed to sets of the nested variants with ivt or ivt like semantics. So selector() gets the flag, and can perform an ivt at each record offset operation. Note that this also implies that the final ivt in assignment() can be skipped.

Note we have homework to do to evaluate VAR passed and read(ln) cases, since they are defacto assignments.