New utilities:
set32 changes the source code to process 32 bit P-machine code.
set64 changes the source code to process 64 bit P-machine code.

Right now, you have to follow whatever bit size the host compiler
has, because the size of integer determines it. Thats the most
efficient way to do it.

We are going to change the configurator to acknoledge this fact.