Fix line endings in configure for ip_pascal.
So with git, things got a bit more complex. The default setting with
git is to convert pulled things to the native format, whatever
that native format is. We deal with this by assuming that bash
scripts are run under unix, and dos scripts are run under DOS/Windows.
That won't be true very long, because bash can run under windows.
Its true now, because bash can't recognize gpc created binaries.
When that gets fixed, we will have to reexamine.