Fix for mark sizing
1. The mark layout was written into constants in the code. I moved the
constants to the machine specific area and created 64 bit versions as
well.

2. Added the mark definitions to set32 and set64 commands.

At this point, the 64 bit version runs a simple program (hello, world of
course) under 64 bit Ubuntu 14.04. The 64 bit GPC required a mod to
work, and although I was able to get a 32 bit version of GPC to run
under that system, chasing down all of the include path violations and
libraries was more than I could bear (just to get a braindamaged 32 bit
on 64 bit os implementation).

Docs, full test runs and other items to follow.