Refactored out the deep nested if-else.
I don't expect anyone to believe me, but the %errorlevel% variable was
getting "stuck", that is, it would have a value of 1 returned by a
failing command ("where"), then it would not be cleared by the next
execution -- of anything.

I scattered various echo %errorlevel% statements around, found it was
being set to 1 by the least nested statement, then stuck after that. On
a hunch, I replaced the nesting by goto statements again.

It immediately started giving the right results. Seems that somehow
cmd.exe is mishandling the if () nesting and causing %errorlevel% to
misbehave.

No, I would probably not believe it either. But it works now. Moving
on...