Fixed several uninitialized variable issues. One of these was an assignment to a variant before establishing the tag, which is a major no no that I suspect we will see a lot more of.
The packing flag was found uninitialized, and that deserves some
explaination. The packing flag is outside of the variants, meaning that
it is generally applicable, but was not being dealt with or initialized
by several structure types, for example scalars.

This might seem to have no meaning, and thus the answer would seem to be
that it needs to go into the variants that need it, ie, records and
arrays.

However, packing is a general purpose concept. Nobody knows what a
packed file is, but that is valid.

Thus the packing flag was initialized everywhere, even if not used.