Error found in free block allocator. When searching for free blocks, and upon finding one big enough, we either allocate the entire block to the request, or break it into a block of exact size and a new block, free, with the remainder.
This was done, but the block was always being set to the request block
size regardless. This corrupts the list if the block does not get broken
in two.

The changes were:

1. Simply allocate the existing block at its size, then allocate both
the request block and the stub block above it if the block can be broken
in two.

2. Added a new error check for length < minimum or > cp (end of heap).
This sanity check will often flag if the list is corrupt. Of course it
takes time...