Overhaul stack space accounting
The cdx and pdx arrays had not been extended to new operators (and thus
were causing uninitialized access errors in the new compiler).

On review, the old method of gauging the stack by the maximum size of
operations was deemed unsatisfactory. maxstack set the largest possible
stack entry, and all entries on stack were assumed to be this length.
This led to a difference in this "rule of thumb" allocation of 1 to 8,
that is, integer vs. sets at 4 bytes to 32 bytes.

The stack space accounting was completely overhauled. The cdx and pdx
arrays now contain actual lengths based on MPB (Machine Parameter Block)
equates. Because typed operands were never broken out by type, the CDX
table now has a two level structure for typed operators, and the CDX
table contains an index into the cdxs (secondary) table, which contains
lengths according to type.

The cdxs table turned out to need only 6 sets of entries, because
several of the type sets were identical and the two level structure
allowed that duplication to be expressed. The result is that the
cdx/cdxs table is similar to the runtime cost of the old system, which
is now accurate down to the byte of storage used.

1. maxstack is removed.

2. cdxs array is added.

3. the cdx and pdx arrays now get integer lengths instead of "logical"
lengths.

4. A new routine, aligns(), aligns addresses specifically for the stack.

5. Compiler errors are now output for the stack going negative at any
time, and for the stack not being zero at the end of the routines. This
serves to catch any issues with the stack allocation system.

6. The new routine mesl() adjusts the stack size by length and all other
routines, including mes() call that. It allocates in absoute length, not
logical lengths.

7. A new routine, mest(), takes a structure pointer, and allocates the
stack by operand type.

8. System routines use the pdx array as real lengths, not logical
lengths.

9. retp bypasses the system with a length of 0.

10. Several operator cases were broken out into special handling.

11. The routine caller, callnonstandard() gets handling for both the
fact that parameters are automatically removed from the stack, as well
as allocating the return value.

12. Casestatement() now removes its selector from the stack when it
jumps to the bottom, and reinstates it at the bottom.

13. Stack accounting now starts at 0 on a new block, not
lcaaftermarkstack. No idea why this was ever correct (including looking
at P4).

14. cdx and pdx array extensively changed and reformatted. cdxs[] added.

The resulting binary was regressed up to self compile, which is down in
this code for other reasons.