Add "byte mode" to compiler
This is a complex change. During self-compile, it becomes quite
problematic if the compiler does not treat byte values specially, it
impacts creation of byte files, and makes reformatting objects in the
interpreter dificult. pint.pas already relies on this capability, as
given by the host, so we extend it to the compiler itself.

A new type postfix is added, "x" ("b" was taken for boolean). This
generally needs to be applied only to loads and stores, because all
operands are extended to full integer type on the stack. The exception
to this is "chkx", where the x refers to the destination, but it is
equivalent to "chki".

An alternative design would have been to add a length parameter to loads
and stores that gave their actual length. This would have added 8, 16,
and 32 (in 64 bit compiler) bit subrange loads/stores. I didn't choose
this because the byte case is functional (can't do byte files without
it, etc.), whereas the others just save space. With the "x" system the
compiler knows two basic sizes, byte, and "N", where N is the length of
a word.

This is a partial commit. There are remaining issues to be fixed.

1. The function isbyte() is added, which makes checking for a byte
lengthed type easy. Note that byte types must fit into 0..255
(unsigned).

2. In simpletype(), subrange types are checked and lengthed for byte.

3. gentypindicator() now recognizes and outputs the byte case.

4. load() normalizes all subranges to their base type after load. This
was formerly done at a low level (factor()), but we need types to carry
their subranges for a bit longer than that. Load is the logical place
for that, because load marks the time in which the value of an operand
is loaded to stack, and thus extended to its full base type.

Note that the base type normalization now extends to any level. Formerly
it was single level, but I find that rule suspect. Subranges of
subranges are certainly possible.

5. Removed alignment in selector().