#26 Character constants outside printable range should not be emitted as char
The compiler is not really supposed to get into what is valid or invalid
in the character set, so I handled it with the chartp[] table, which
already exists. Now, ldcc can take its argument either as a quoted
character or a number, in which case it converts that to the equivalent
character code. To keep the intermediate readable, I let the chartp[]
determine what is printable and not, and what form to put the character
in, number or quoted character.

I didn't extend this same courtesy to strings, because they come from
the user, and it is up to the user to keep them printable. This will
have to change in p6.