Added ability to output options to intermediate
pcom now can output options to the intermediate, and these are read in
pint and stored. Any number of options can appear on an intermediate
output line. pcom outputs a full line of options according to the
compiler defaults at the start of the intermediate, then any number of
single option set lines are output after as user options are parsed from
the source.

This couples the options from front end to back end. At present, there
are no options that cross boundaries, but this enhancement will change
that.

Options are tracked in an array of 'a'..'z', and we transmit all
options, not just the currently valid ones. This allows second stage
only options. The original pcom option variables are still used and
shadowed, which was more convienent and memnemonic than getting rid of
the original variables. I suggest this be maintained.