Change ticket #45 to issue compiler fault on miss to the cdxs table
cdx is an overloaded table (I am actually famous for that). It contains
stack offsets for a series of instructions, but the ones that need to be
expanded to a per-type definition go to secondary tables by context,
that is, it is entirely up to the code to redirect to the correct table.

The failure indicates to me that a standard table entry for cdx got
misinterpreted as an overloaded table entry. That needs finding.