Add P5 Basic
P5 Basic is an old (1994) Full Basic interpreter developed under
IP Pascal. It is entered as a test program for P5 here, but could
be shaped up into a full programming facility if that is wanted.

Added the main interpreter, a series of Basic program files to
run and test with, documentation, and test files to allow P5 to
run a batch regress on it.