FPC related fixups
1. Fixed FPC warnings for unused variables.

2. Turned off warnings in FPC compile.

FPC flags things like unreachable code, of which there are a lot
of in P5 (horrors!!!). These are the options like dodmplex that
are used to remove optional parts of the code.

FPC does not give me the option to turn individual warnings off.
I notice the same thing was done with GPC.