Upgrade to linux configure scripts
1. Fixed make_flip script.
2. Reoriented the configure script to search for available Pascal
compilers in a priority order. This matches what the configure.bat
script does.
3. Added FPC configure target.

The ability to search for an available compiler is an important
feature of the configure scripts and matches what GNU system
does. A primer on this is:

https://en.wikipedia.org/wiki/GNU_Build_System

The bottom line is that, although we don't do the same complex
configs that the autotools system does, we try to achieve
roughly the same result. Meaning users have an expectation of
what the system does.