Removed FPC from configure scripts and code
FPC claims there was a fix and that would end up in a release. Its been
more than a year, still no fix. When and if they do release a fix, I can
bring the support back here.

Note I also removed the p5c support, this really belongs in the p5c
project, since it is undocumented and unsupported in this archive.