Partial recommit: convergence point source label generation
Changes to be committed:
      modified:   source/pcom.pas

1. Change ':' source line marker generation to simple routine.

2. Add convergence markers to ifstatement().

The convergence fix is to solve the issue where, when we mark source
line passage in the intermediate, this may not give an accurate account
when executing of where we are in the source vs. the intermediate
execution. The reason for this is the code may just jump over the
source markers. This will then leave the source line at whatever we
jumped from.

The solution to this is to output (possibly redundant) markers after
a convergence point in the code. That way, every jump will then encounter
a source marker.

This works because Pascal-P5 does not care if the source line marker
is redundant. In Pascal-P6, which converts source line markers to ranges
of code in a tracking array, the redundant markers have to be ignored.