Recommit last undefined tag fix; adust client programs; add regress report
Changes to be committed:
      modified:   basic/basic.pas

1. Added flag to turn off undefined tagfield check.

      new file:   regress_report.txt

1. Added regression reporting from Pascal-P6.

      modified:   sample_programs/pascals.pas

1. Added flag to turn off undefined tagfield check.

      modified:   source/pcom.pas

1. Added options comment.

2. Added back undefined tagfield checking.

Regression reporting is a Pascal-P6 feature. The regression run creates
a report of all tests and their results, the time of run, and an md5 sum
of all the tooling and scripts used to perform the regression.

This essentially "certifies" a checkin, since it proves (or disproves)
that the code worked against all current tests, along with the ability to
verify all the tooling that was used.

This is presently a PAT only test. We will add PRT. A versioned release
also gets self-compile, but this is too time consuming for regular checkins.