Backport of alias fix from Pascal-P6
The alias fix enables recognition of defining point errors from ISO 7185.
This is where a program references identifiers from a surrounding block,
but then defines the same symbol in the same block. This leads identifier
references of the same name, in the same block, but leading to different
blocks. This clearly violates ISO 7185.

1. Removed references to products outside of P5.

1. This is an error, and will be reverted.

1. Added alias class to ids.

2. Added define flag to displays, marks if the block will perform definitions.
For example, procedures are defining blocks and records are not (yes, really).

3. Changed putname to process aliases.

4. Changed putname to recycle according to type. This fix is unrelated to aliases,
but it causes an error in tag checked versions of Pascal.

5. Added defining point error.

6. Changed enterid() to process error according to if the id is an alias, and
also to skip over aliases.

7. Changed serachidnenm to handle aliases.

8. Searchid() changed to create aliases.

9. Changed printtables() to print alias entries.

10. Checkrefs() changed to handle aliases.

11. Initialize the define flag for blocks at various block starts.

12. Removed old mark/release comments, obsolete.

1. Various PRT updates.