Self compile fixes
Changes to be committed:
      modified:   bin/pascpp.bat
      modified:   regress_report.txt
      modified:   source/pcom.pas

1. All of the changed statements violated the rule that the tagfield must
be set before the variant is accessed. This was just never checked.

The new style is that you allocate the entry with new(), then immediately
assign to the tags used in the new(). Then whatever follows will be valid.

I'm sure there are many such errors.