Fixes for self compile
Changes to be committed:
      modified:   bin/cpcoms.bat

1. Final form for pcom.pas self compile.

      modified:   regress_report.txt
      modified:   source/pcom.pas

1. putcst() now recycles by type. Note that using a general dispose
after a type specific one does not hurt, since the heap manager knows
the size of the block and recycles it regardless. However, this is
strict checking of the ISO 7185 rules.

2. putstc() recycles by type.

cpcoms works, so self compile of pcom.pas works complete. Next is
self compile of pint.pas.