Prevent exponent overflow
Changes to be committed:
      modified:   regress_report.txt
      modified:   source/mpb16.inc
      modified:   source/mpb32.inc
      modified:   source/mpb64.inc

1. Estabished maxexp constant to set limit of exponent.

      modified:   source/pcom.pas

1. If the exponent limit is exceeded, we flag an error (if one is not
already flagged) and clear the exponent to zero. Note that the number
is limited by maxint on input, but checked against (what we assume to
be a much lower) maxexp after.

Ideally, the number, even too large, should be carried to the backend,
which was Wirth/Amman's original plan, but this causes problems reading
the number there.