Fix for PRT test 1854: Self referencing type. Same type but outter block.
 Changes to be committed:
       modified:   regress_report.txt
       modified:   source/pcom.pas

1. Added defined flag to identifiers.

2. In ininam(), always set defined to true.

3. Add error for type reference incomplete.

4. In searchid(), show error if referenced entry is incomplete.

5. In typedeclaration(), we change the behavior to enter the type id
before the type equivalent is parsed, but flag the entry as incomplete.
Once the type is complete, then the defined flag is turned on again.

This is an overly broad solution to the 1854 issue, because it affects
all ids, not just type. However, it was cheaper than finding and fixing
all type id references.