Enhancements to ISO 7185 PAT.
1. Expanded the Impleentation defined section using output to a temp file,
then readback an parse to determine several parameters. One of these is
also listed as implementation dependent (page). The rest of the implementation
dependent items from the standard, mostly order of execution dependencies,
are testable, but I don't feel are wise to include in a ISO 1785 test because
they could change depending on an optimizing compiler. The last item, 6.10
"binding of program parameters to external entities" is testable, but belongs
in an external test due to it's completely implementation dependent nature.

2. Performed integer testing on buffer variables. Many more tests like this
need to be added.