Put in new algorithim to detect define after use.
Switched to using the Arthur Sale algorithim. Unfortuately, the
symbol table routines got even more messy, so looking forward to
a code cleanup.

 Changes to be committed:
       modified:   regress_report.txt
       modified:   source/pcom.pas

1. Added lastuse field to identifier.

2. Added counter for scopes, scopen.

3. ininam now sets up the lastuse field.

4. Introduced a new general search routine which is not qualified
by a type class, searchidlvs() (search id levels).

5. enterid() uses searchidlvs() to find downlevel ids. Then we check
if the lastuse field is not less than the current scope count. If so,
a use before define error follows. enterid() also initalizes the
lastuse field for the new id.

6. searchidne() sets the lastuse field from the scope count.

7. (various places) the scopen count is incremented each time a block
start occurs. Note that it is a linear count of the scopes opened.

       modified:   standard_tests/iso7185pat.pas

1. This is an error and will be corrected.

       modified:   standard_tests/iso7185prt.rpt
       modified:   standard_tests/iso7185prt1915.ecp
       modified:   standard_tests/iso7185prt1917.ecp

1. Reran the PRT and fixed the changed compare files.