Updated docs; refactored with tracking; fixed with bug.
Was bug in with tracking with respect to gotos.

 Changes to be committed:
       modified:   doc/the_p5_compiler.docx

1. Added section on submitting bugs.

       modified:   regress_report.txt
       modified:   source/mpb16.inc
       modified:   source/mpb32.inc

1. Added with base counter to mark.

       modified:   source/pcom.pas

1. Added "with" tracking entries.

2. Added "with" tracking stack.

3. Added pshwth() and popwth() routines, which push and pop with stack
levels.

4. On branch taken, goto compares the target label with statement
level to the ones on the with stack, and traverses the stack until
there is a match. While doing that, it issues with block ends.

5. In casestatement, there was a bug where we were doing a putcas()
(dispose case entry) on the entry opened in the current with. The
solution was to move it outside the with, which was completely
equivalent. This was the bug that set off all of the rest of this,
and it was found by running pcom.pas on the virtual machine to
take advantage of advanced error checking.

6. Added push and pop with stack to withstatement(). Otherwise it
still uses counts the previous behavior. The stack is just for
reference in a goto statement.

       modified:   source/pint.pas

1. Added with stack counter to withenter() and withexit().

2. The with block counter is stored in the mark.

3. On ipj, we retrive the target with count, then cut the stack until
that number is seen. If something goes wrong, the stack will empty
out and give an error.

4. Added errors at the end of run for with and VAR stacks out of
balance.

Need to add recycle balance error for with tracking entries.